# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Financial Testing Sandbox API",
    "version": "1.0.0",
    "description": "Complete API documentation including status, users, utilities, and auth routes."
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root path info",
        "tags": [
          "Utilities"
        ],
        "responses": {
          "200": {
            "description": "Server is running."
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Returns application health",
        "tags": [
          "Status"
        ],
        "responses": {
          "200": {
            "description": "App is healthy"
          }
        }
      }
    },
    "/api/status": {
      "get": {
        "summary": "Returns backend status",
        "tags": [
          "Status"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Retrieve a list of test users",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of users"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/login": {
      "get": {
        "summary": "Render login form",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Login form HTML"
          }
        }
      },
      "post": {
        "summary": "Authenticate a user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "johndoe"
                  },
                  "password": {
                    "type": "string",
                    "example": "secure123"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/register": {
      "get": {
        "summary": "Render registration form",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Registration form HTML"
          }
        }
      },
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "middleName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "jobTitle": {
                    "type": "string"
                  },
                  "mobile": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "email",
                  "password",
                  "firstName",
                  "lastName",
                  "jobTitle"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "409": {
            "description": "User already exists"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}